refactor(workers): clarify Zendesk resource layout
